requirements.txt
numpy

# only needed for testing/debugging
pillow
pyrr